remove v2 files, move changes to original files
