+++ model_code/evaluation.py
@@ -1,10 +1,11 @@
 from constants import *
-from preprocess import decode_class
+import os
+import pickle
+from preprocess import key_to_filename
 
 class Results():
-    def __init__(self, models, split_phase_types):
-        self.models = models
-        self.split_phase_types = split_phase_types
+    def __init__(self, model_path):
+        self.model_path = model_path
 
         self.class_corrects = dict()
         self.class_totals = dict()
@@ -16,50 +17,44 @@ class Results():
 
     def __repr__(self):
         output = f"Complete Correct: {self.all_correct}\nComplete Total: {self.all_total}\nComplete Accuracy: {(100 * self.all_accuracy):.2f}%"
-        if self.split_phase_types:
-            for phase_type in self.class_accuracies.keys():
-                output += f"\nClass Correct ({phase_type}): {(self.class_corrects[phase_type])}\n"
-                output += f"Class Total ({phase_type}): {(self.class_totals[phase_type])}\n"
-                output += f"Class Accuracy ({phase_type}): {(100 * self.class_accuracies[phase_type]):.2f}%\n"
+        for model_type in self.class_accuracies.keys():
+            output += f"\nClass Correct ({model_type}): {(self.class_corrects[model_type])}\n"
+            output += f"Class Total ({model_type}): {(self.class_totals[model_type])}\n"
+            output += f"Class Accuracy ({model_type}): {(100 * self.class_accuracies[model_type]):.2f}%\n"
         return output
 
     def evaluate(self, test_dict):
-        for phase_type, data in test_dict.items():
-            print(f"Predicting for phases {phase_type}")
-            pred_labels = self.models[phase_type].predict(data[0])
-            pred_orders = map(decode_class, pred_labels)
-            true_orders = map(decode_class, data[1])
+        for model_type, data in test_dict.items():
+            print(f"Predicting for key {model_type}")
 
             class_correct = 0
             class_total = 0
+            true_orders = data[1]
 
-            for pred, true in zip(pred_orders, true_orders):
-                correct, total = order_accuracy(pred, true)
-                class_correct += correct
-                class_total += total
+            file_path = os.path.join(self.model_path, key_to_filename(model_type))
+            if os.path.exists(file_path):
+                pred_orders = []
+                with open(file_path, 'rb') as model_file:
+                    model = pickle.load(model_file)
+                    pred_orders = model.predict(data[0])
 
-            self.class_corrects[phase_type] = class_correct
-            self.class_totals[phase_type] = class_total
-            self.class_accuracies[phase_type] = class_correct / class_total
+                for pred, true in zip(pred_orders, true_orders):
+                    if (pred == true):
+                        class_correct += 1
+                    class_total += 1
+            else:
+                print(f"Model not found | key: {model_type}")
+                class_total += len(true_orders)
+
+            self.class_corrects[model_type] = class_correct
+            self.class_totals[model_type] = class_total
+            self.class_accuracies[model_type] = class_correct / class_total
             self.all_correct += class_correct
             self.all_total += class_total
 
         self.all_accuracy = self.all_correct / self.all_total
 
-def order_accuracy(predicted, true):
-    correct = 0
-    total = 0
-
-    for i, power in enumerate(POWERS):
-        for order in predicted[i]:
-            correct += order in true[i]
-            total += 1
-        for order in true[i]:
-            correct += order in predicted[i]
-            total += 1
-    return correct, total
-
-def evaluate_model(test_dict, models, split_phase_types=False):
-    results = Results(models, split_phase_types)
+def evaluate_model(test_dict, model_path):
+    results = Results(model_path)
     results.evaluate(test_dict)
     return results
